pyproject.toml
[project]
name = "jax-cmb-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "blackjax~=1.0.0",
    "candl-data",
    "candl-like>=2.2.0",
    "clipy-like>=0.15",
    "cosmopower-jax==0.5.7",
    "getdist>=1.7.4",
    "ipykernel>=7.1.0",
    "ipympl>=0.9.8",
    "ipywidgets>=8.1.8",
    "jax==0.4.16",
    "jaxlib==0.4.16",
    "jupyter>=1.1.1",
    "matplotlib>=3.5.0",
    "numpy<2",
    "pyyaml>=6.0",
    "scipy<1.13",
    "seaborn>=0.13.2",
    "tensorflow<2.14",
    "tqdm>=4.67.1",
    "typing-extensions==4.15",
]

[dependency-groups]
dev = []

[tool.uv.sources]
candl-data = { path = "../candl_data" }
cosmopower-jax = { path = "../cosmopower-jax" }
